Add guestfish script to create single boot image
src/boot_disk.sh:
Add guestfish script to create single boot image, which can later on be
used to construct a combined init image, as well as RAUC update bundles.
NOTE: It is currently still unclear whether block devices in update
bundles are supposed to carry a filesystem directly on a block device or
on a partition in a block device, so we offer both (commented).