Add guestfish script to create a combined image
src/combined_disk.sh:
Add guestfish script to create a combined image, which serves as initial
image for a target system.
NOTE: It is currently still unclear whether block devices in update
bundles are supposed to carry a filesystem directly on a block device or
on a partition in a block device, so we offer both (commented).